Struct rust_htslib::bam::pileup::Alignment

pub struct Alignment<'a> { /* fields omitted */ }

An aligned read in a pileup.

Methods

impl<'a> Alignment<'a>

fn new(inner: &'a bam_pileup1_t) -> Self

fn qpos(&self) -> Option<usize>

Position within the read. None if either is_del or is_refskip.

fn indel(&self) -> Indel

Insertion, deletion (with length) if indel starts at next base or None otherwise.

fn is_del(&self) -> bool

Whether there is a deletion in the alignment at this position.

fn is_head(&self) -> bool

Whether the alignment starts at this position.

fn is_tail(&self) -> bool

Whether the alignment ends at this position.

fn is_refskip(&self) -> bool

Whether this position is marked as refskip in the CIGAR string.

fn record(&self) -> Record

The corresponding record.