Struct rust_htslib::bcf::record::Record

pub struct Record {
    pub inner: *mut bcf1_t,
    pub header: *mut bcf_hdr_t,
    // some fields omitted
}

Fields

inner

header

Methods

impl Record

fn new() -> Self

fn rid(&self) -> Option<u32>

fn pos(&self) -> u32

fn set_pos(&mut self, pos: i32)

fn alleles(&self) -> Vec<&[u8]>

fn qual(&self) -> f32

fn set_qual(&mut self, qual: f32)

fn info<'a>(&'a mut self, tag: &'a [u8]) -> Info

Get the value of the given info tag.

fn sample_count(&self) -> u32

fn allele_count(&self) -> u16

fn format<'a>(&'a mut self, tag: &'a [u8]) -> Format

Get the value of the given format tag for each sample.

fn push_format_integer(&mut self, tag: &[u8], data: &[i32]) -> Result<(), ()>

Add an integer format tag. Data is a flattened two-dimensional array. The first dimension contains one array for each sample. Returns error if tag is not present in header.

fn push_format_float(&mut self, tag: &[u8], data: &[f32]) -> Result<(), ()>

Add a float format tag. Data is a flattened two-dimensional array. The first dimension contains one array for each sample. Returns error if tag is not present in header.

fn trim_alleles(&mut self) -> Result<(), ()>

Remove unused alleles.